# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  id: number;
  name: string;
  hash: string;
  createdAt: string;
}

export interface DocumentInput {
  name: string;
  hash: string;
}

export interface ScaleConfig {
  documentId: number;
  isSet: boolean;
  pixelsPerUnit: number;
  unit: string;
  realWorldUnit: string;
}

export interface ScaleInput {
  isSet: boolean;
  pixelsPerUnit: number;
  unit: string;
  realWorldUnit: string;
}

export type AnnotationType = typeof AnnotationType[keyof typeof AnnotationType];


export const AnnotationType = {
  highlight: 'highlight',
  note: 'note',
  text: 'text',
} as const;

export type AnnotationFabricData = { [key: string]: unknown };

export interface Annotation {
  id: string;
  documentId: number;
  pageNumber: number;
  type: AnnotationType;
  fabricData: AnnotationFabricData;
  createdAt: string;
}

export type AnnotationInputType = typeof AnnotationInputType[keyof typeof AnnotationInputType];


export const AnnotationInputType = {
  highlight: 'highlight',
  note: 'note',
  text: 'text',
} as const;

export type AnnotationInputFabricData = { [key: string]: unknown };

export interface AnnotationInput {
  id: string;
  pageNumber: number;
  type: AnnotationInputType;
  fabricData: AnnotationInputFabricData;
}

export type MeasurementType = typeof MeasurementType[keyof typeof MeasurementType];


export const MeasurementType = {
  distance: 'distance',
  area: 'area',
} as const;

export type MeasurementPointsItem = { [key: string]: unknown };

export type MeasurementFabricData = { [key: string]: unknown };

export interface Measurement {
  id: string;
  documentId: number;
  pageNumber: number;
  type: MeasurementType;
  label: string;
  realWorldValue: number;
  unit: string;
  points: MeasurementPointsItem[];
  fabricData: MeasurementFabricData;
  createdAt: string;
}

export type MeasurementInputType = typeof MeasurementInputType[keyof typeof MeasurementInputType];


export const MeasurementInputType = {
  distance: 'distance',
  area: 'area',
} as const;

export type MeasurementInputPointsItem = { [key: string]: unknown };

export type MeasurementInputFabricData = { [key: string]: unknown };

export interface MeasurementInput {
  id: string;
  pageNumber: number;
  type: MeasurementInputType;
  label: string;
  realWorldValue: number;
  unit: string;
  points: MeasurementInputPointsItem[];
  fabricData: MeasurementInputFabricData;
}

export interface Share {
  id: number;
  token: string;
  documentId: number;
  createdAt: string;
  /** @nullable */
  expiresAt: string | null;
}

export interface ShareInput {
  documentId: number;
}

export interface SharePayload {
  document: Document;
  scale: ScaleConfig;
  annotations: Annotation[];
  measurements: Measurement[];
  share: Share;
}