# lib/api-zod/src/generated/types/annotation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationFabricData } from './annotationFabricData';
import type { AnnotationType } from './annotationType';

export interface Annotation {
  id: string;
  documentId: number;
  pageNumber: number;
  type: AnnotationType;
  fabricData: AnnotationFabricData;
  createdAt: string;
}