# lib/api-zod/src/generated/types/annotationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationInputFabricData } from './annotationInputFabricData';
import type { AnnotationInputType } from './annotationInputType';

export interface AnnotationInput {
  id: string;
  pageNumber: number;
  type: AnnotationInputType;
  fabricData: AnnotationInputFabricData;
}