# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  name: string;
  hash: string;
  createdAt: string;
}