# lib/api-zod/src/generated/types/measurement.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementFabricData } from './measurementFabricData';
import type { MeasurementPointsItem } from './measurementPointsItem';
import type { MeasurementType } from './measurementType';

export interface Measurement {
  id: string;
  documentId: number;
  pageNumber: number;
  type: MeasurementType;
  label: string;
  realWorldValue: number;
  unit: string;
  points: MeasurementPointsItem[];
  fabricData: MeasurementFabricData;
  createdAt: string;
}