# lib/api-zod/src/generated/types/measurementInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementInputFabricData } from './measurementInputFabricData';
import type { MeasurementInputPointsItem } from './measurementInputPointsItem';
import type { MeasurementInputType } from './measurementInputType';

export interface MeasurementInput {
  id: string;
  pageNumber: number;
  type: MeasurementInputType;
  label: string;
  realWorldValue: number;
  unit: string;
  points: MeasurementInputPointsItem[];
  fabricData: MeasurementInputFabricData;
}