# lib/api-zod/src/generated/types/share.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Share {
  id: number;
  token: string;
  documentId: number;
  createdAt: string;
  /** @nullable */
  expiresAt: string | null;
}