# lib/api-zod/src/generated/types/sharePayload.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Annotation } from './annotation';
import type { Document } from './document';
import type { Measurement } from './measurement';
import type { ScaleConfig } from './scaleConfig';
import type { Share } from './share';

export interface SharePayload {
  document: Document;
  scale: ScaleConfig;
  annotations: Annotation[];
  measurements: Measurement[];
  share: Share;
}